Extracted the download implementation from FringeEventDownloader into DownloadHelper
